First version of scribo library along with sample application.
scribo enables users to route the debug logs to the ADB console
as well as write it to a file on the SDCard for later use. The
DebugHelper class in the scribo library provides various types
of functions to achieve the following:
- Set the name of the file in which the logs need to be recorded.
- Override the default log categories provided by scribo with the
  ones which are more relevant for the users.
- Enable/Disable any of the log categories dynamically. (Except
  GENERAL all other category logs are disabled by default)

A simple sample application is provided to demonstrate the usage
of scribo.